work towards fixing tests
